# public/openapi.yaml
openapi: 3.1.0
info:
  title: ChronoChimp API
  version: v1.0.0
  description: API for managing tasks and user settings in the ChronoChimp application.
servers:
  - url: /api
    description: Main API server
components:
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized: Invalid or missing authentication token"
              error:
                type: string
                example: "unauthorized"
              statusCode:
                type: integer
                example: 401
    ForbiddenError:
      description: Forbidden - Insufficient permissions.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Forbidden"
    NotFoundError:
      description: Not Found - The requested resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found."
    ValidationError:
      description: Bad Request - Invalid input data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    UserRole:
      type: string
      enum: [Admin, Editor, Viewer]
      example: Viewer
    User: # For general user display, not for sensitive creation/update
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "user-abc-123"
        name:
          type: string
          example: "Jane Doe"
        email:
          type: string
          format: email
          example: "jane.doe@example.com"
        role:
          $ref: '#/components/schemas/UserRole'
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: "https://placehold.co/100x100.png"
        isTwoFactorEnabled:
          type: boolean
          description: "Indicates if 2FA is enabled for the user."
          example: false
        emailNotificationsEnabled:
          type: boolean
          description: "User preference for email notifications."
          example: true
        inAppNotificationsEnabled:
          type: boolean
          description: "User preference for in-app notifications."
          example: true
        smtpHost:
          type: string
          nullable: true
          example: "smtp.example.com"
        smtpPort:
          type: integer
          nullable: true
          example: 587
        smtpEncryption:
          type: string
          enum: [none, ssl, starttls]
          nullable: true
          example: "starttls"
        smtpUsername:
          type: string
          nullable: true
          example: "user@example.com"
        smtpSendFrom:
          type: string
          format: email
          nullable: true
          description: "Default 'send from' email for SMTP."
          example: "noreply@example.com"
        joinedDate:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
...
    UserUpdateRequest: # Admin updating a user
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
          nullable: true
          description: "Optional. New password for the user. Will be hashed if provided."
        role:
          $ref: '#/components/schemas/UserRole'
        avatarUrl:
          type: string
          format: uri # Or allow data URI
          nullable: true
        isTwoFactorEnabled:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid input."
        details: # For Zod validation errors
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email address."]
            password: ["Password must be at least 6 characters."]
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: "Operation successful."
  securitySchemes:
    bearerAuth: # Can be used to denote JWT authentication
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /admin/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: ID of the user to operate on
        schema:
          type: string
          format: uuid
    get:
      summary: Get a specific user's details (Admin only)
      tags: [Admin - User Management]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Detailed information about the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update an existing user's details (Admin only)
      tags: [Admin - User Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409': # If email is changed to one that's already in use
          description: Email address already in use.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Email address already in use by another account."
    delete:
      summary: Delete a user (Admin only)
      tags: [Admin - User Management]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/invites:
    post:
      summary: Create a user invite (Admin only)
      tags: [Admin - User Management, User Invites]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                role:
                  $ref: '#/components/schemas/UserRole'
                  default: "Viewer"
      responses:
        '201':
          description: User invite successfully generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
                properties:
                  message:
                    example: "Invite successfully generated. If SMTP is configured, an email has been sent to user@example.com. Otherwise, check server logs for invite details if fallback logging is enabled in emailService.ts."
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Email already registered or invite already pending.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "This email address is already registered."
  # End of API paths